+++ core/src/main/java/arez/spy/ObserveCompletedEvent.java
@@ -5,16 +5,16 @@ import java.util.Objects;
 import javax.annotation.Nonnull;
 
 /**
- * Notification when Observer completes reaction.
+ * Notification when Observer completes method being observed.
  */
-public final class ReactionCompletedEvent
+public final class ObserveCompletedEvent
   implements SerializableEvent
 {
   @Nonnull
   private final ObserverInfo _observer;
   private final int _duration;
 
-  public ReactionCompletedEvent( @Nonnull final ObserverInfo observer, final int duration )
+  public ObserveCompletedEvent( @Nonnull final ObserverInfo observer, final int duration )
   {
     assert duration >= 0;
     _observer = Objects.requireNonNull( observer );
@@ -38,7 +38,7 @@ public final class ReactionCompletedEvent
   @Override
   public void toMap( @Nonnull final Map<String, Object> map )
   {
-    map.put( "type", "ReactionCompleted" );
+    map.put( "type", "ObserveCompleted" );
     map.put( "name", getObserver().getName() );
     map.put( "duration", getDuration() );
   }